pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cfeintact"
version = "1.26.1"
description = "A bioinformatics tool for detecting defects in HIV-1 proviral sequences."
requires-python = ">=3.11,<3.13"
# readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Imogen Wright", email = "imogen@hyraxbio.co.za" },
    { name = "British Columbia Centre for Excellence in HIV/AIDS", email = "vmysak@bccfe.ca" },
]
dependencies = [
    "biopython>=1.83",
    "click>=8.0",
    "scipy>=1.1",
    "numpy==2.4.1",
    "aligntools==1.2.2",
    "mappy>=2.24",
]

[project.optional-dependencies]
test = [
    # Dependencies required for running the test suite
    "pytest",
    "pytest-cov",
    "requests",
]
dev = [
    # Dependencies required for development (linting, type checking, etc.)
    "flake8",
    "mypy",
    "types-requests",
]

[project.scripts]
cfeintact = "cfeintact.main:cli"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = [ "src/cfeintact" ]

[tool.pytest.ini_options]
pythonpath = "src"
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "overnight: for tests that take hours to complete (deselect with '-m \"not overnight\"')",
  "serial"
]

[tool.mypy]
files = ["src", "tests"]
strict_optional = true
allow_untyped_calls = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
strict = true

[[tool.mypy.overrides]]
module = ["scipy.*", "pytest.*", "mappy"]
ignore_missing_imports = true
